+++ x/bank/msgs_test.go
@@ -13,8 +13,8 @@ func TestNewMsgSend(t *testing.T) {}
 
 func TestMsgSendType(t *testing.T) {
 	// Construct a MsgSend
-	addr1 := sdk.Address([]byte("input"))
-	addr2 := sdk.Address([]byte("output"))
+	addr1 := sdk.AccAddress([]byte("input"))
+	addr2 := sdk.AccAddress([]byte("output"))
 	coins := sdk.Coins{sdk.NewCoin("atom", 10)}
 	var msg = MsgSend{
 		Inputs:  []Input{NewInput(addr1, coins)},
@@ -26,12 +26,12 @@ func TestMsgSendType(t *testing.T) {
 }
 
 func TestInputValidation(t *testing.T) {
-	addr1 := sdk.Address([]byte{1, 2})
-	addr2 := sdk.Address([]byte{7, 8})
+	addr1 := sdk.AccAddress([]byte{1, 2})
+	addr2 := sdk.AccAddress([]byte{7, 8})
 	someCoins := sdk.Coins{sdk.NewCoin("atom", 123)}
 	multiCoins := sdk.Coins{sdk.NewCoin("atom", 123), sdk.NewCoin("eth", 20)}
 
-	var emptyAddr sdk.Address
+	var emptyAddr sdk.AccAddress
 	emptyCoins := sdk.Coins{}
 	emptyCoins2 := sdk.Coins{sdk.NewCoin("eth", 0)}
 	someEmptyCoins := sdk.Coins{sdk.NewCoin("eth", 10), sdk.NewCoin("atom", 0)}
@@ -68,12 +68,12 @@ func TestInputValidation(t *testing.T) {
 }
 
 func TestOutputValidation(t *testing.T) {
-	addr1 := sdk.Address([]byte{1, 2})
-	addr2 := sdk.Address([]byte{7, 8})
+	addr1 := sdk.AccAddress([]byte{1, 2})
+	addr2 := sdk.AccAddress([]byte{7, 8})
 	someCoins := sdk.Coins{sdk.NewCoin("atom", 123)}
 	multiCoins := sdk.Coins{sdk.NewCoin("atom", 123), sdk.NewCoin("eth", 20)}
 
-	var emptyAddr sdk.Address
+	var emptyAddr sdk.AccAddress
 	emptyCoins := sdk.Coins{}
 	emptyCoins2 := sdk.Coins{sdk.NewCoin("eth", 0)}
 	someEmptyCoins := sdk.Coins{sdk.NewCoin("eth", 10), sdk.NewCoin("atom", 0)}
@@ -110,8 +110,8 @@ func TestOutputValidation(t *testing.T) {
 }
 
 func TestMsgSendValidation(t *testing.T) {
-	addr1 := sdk.Address([]byte{1, 2})
-	addr2 := sdk.Address([]byte{7, 8})
+	addr1 := sdk.AccAddress([]byte{1, 2})
+	addr2 := sdk.AccAddress([]byte{7, 8})
 	atom123 := sdk.Coins{sdk.NewCoin("atom", 123)}
 	atom124 := sdk.Coins{sdk.NewCoin("atom", 124)}
 	eth123 := sdk.Coins{sdk.NewCoin("eth", 123)}
@@ -124,7 +124,7 @@ func TestMsgSendValidation(t *testing.T) {
 	output3 := NewOutput(addr2, eth123)
 	outputMulti := NewOutput(addr2, atom123eth123)
 
-	var emptyAddr sdk.Address
+	var emptyAddr sdk.AccAddress
 
 	cases := []struct {
 		valid bool
@@ -178,8 +178,8 @@ func TestMsgSendValidation(t *testing.T) {
 }
 
 func TestMsgSendGetSignBytes(t *testing.T) {
-	addr1 := sdk.Address([]byte("input"))
-	addr2 := sdk.Address([]byte("output"))
+	addr1 := sdk.AccAddress([]byte("input"))
+	addr2 := sdk.AccAddress([]byte("output"))
 	coins := sdk.Coins{sdk.NewCoin("atom", 10)}
 	var msg = MsgSend{
 		Inputs:  []Input{NewInput(addr1, coins)},
@@ -194,9 +194,9 @@ func TestMsgSendGetSignBytes(t *testing.T) {
 func TestMsgSendGetSigners(t *testing.T) {
 	var msg = MsgSend{
 		Inputs: []Input{
-			NewInput(sdk.Address([]byte("input1")), nil),
-			NewInput(sdk.Address([]byte("input2")), nil),
-			NewInput(sdk.Address([]byte("input3")), nil),
+			NewInput(sdk.AccAddress([]byte("input1")), nil),
+			NewInput(sdk.AccAddress([]byte("input2")), nil),
+			NewInput(sdk.AccAddress([]byte("input3")), nil),
 		},
 	}
 	res := msg.GetSigners()
@@ -207,7 +207,7 @@ func TestMsgSendGetSigners(t *testing.T) {
 /*
 // what to do w/ this test?
 func TestMsgSendSigners(t *testing.T) {
-	signers := []sdk.Address{
+	signers := []sdk.AccAddress{
 		{1, 2, 3},
 		{4, 5, 6},
 		{7, 8, 9},
@@ -233,10 +233,10 @@ func TestNewMsgIssue(t *testing.T) {
 
 func TestMsgIssueType(t *testing.T) {
 	// Construct an MsgIssue
-	addr := sdk.Address([]byte("loan-from-bank"))
+	addr := sdk.AccAddress([]byte("loan-from-bank"))
 	coins := sdk.Coins{sdk.NewCoin("atom", 10)}
 	var msg = MsgIssue{
-		Banker:  sdk.Address([]byte("input")),
+		Banker:  sdk.AccAddress([]byte("input")),
 		Outputs: []Output{NewOutput(addr, coins)},
 	}
 
@@ -249,10 +249,10 @@ func TestMsgIssueValidation(t *testing.T) {
 }
 
 func TestMsgIssueGetSignBytes(t *testing.T) {
-	addr := sdk.Address([]byte("loan-from-bank"))
+	addr := sdk.AccAddress([]byte("loan-from-bank"))
 	coins := sdk.Coins{sdk.NewCoin("atom", 10)}
 	var msg = MsgIssue{
-		Banker:  sdk.Address([]byte("input")),
+		Banker:  sdk.AccAddress([]byte("input")),
 		Outputs: []Output{NewOutput(addr, coins)},
 	}
 	res := msg.GetSignBytes()
@@ -263,7 +263,7 @@ func TestMsgIssueGetSignBytes(t *testing.T) {
 
 func TestMsgIssueGetSigners(t *testing.T) {
 	var msg = MsgIssue{
-		Banker: sdk.Address([]byte("onlyone")),
+		Banker: sdk.AccAddress([]byte("onlyone")),
 	}
 	res := msg.GetSigners()
 	require.Equal(t, fmt.Sprintf("%v", res), "[6F6E6C796F6E65]")