+++ x/ibc/ibc_test.go
@@ -28,11 +28,11 @@ func defaultContext(key sdk.StoreKey) sdk.Context {
 	return ctx
 }
 
-func newAddress() crypto.Address {
-	return crypto.GenPrivKeyEd25519().PubKey().Address()
+func newAddress() sdk.AccAddress {
+	return sdk.AccAddress(crypto.GenPrivKeyEd25519().PubKey().Address())
 }
 
-func getCoins(ck bank.Keeper, ctx sdk.Context, addr crypto.Address) (sdk.Coins, sdk.Error) {
+func getCoins(ck bank.Keeper, ctx sdk.Context, addr sdk.AccAddress) (sdk.Coins, sdk.Error) {
 	zero := sdk.Coins(nil)
 	coins, _, err := ck.AddCoins(ctx, addr, zero)
 	return coins, err