+++ x/slashing/signing_info_test.go
@@ -4,11 +4,13 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 func TestGetSetValidatorSigningInfo(t *testing.T) {
 	ctx, _, _, keeper := createTestInput(t)
-	info, found := keeper.getValidatorSigningInfo(ctx, addrs[0])
+	info, found := keeper.getValidatorSigningInfo(ctx, sdk.ValAddress(addrs[0]))
 	require.False(t, found)
 	newInfo := ValidatorSigningInfo{
 		StartHeight:         int64(4),
@@ -16,8 +18,8 @@ func TestGetSetValidatorSigningInfo(t *testing.T) {
 		JailedUntil:         int64(2),
 		SignedBlocksCounter: int64(10),
 	}
-	keeper.setValidatorSigningInfo(ctx, addrs[0], newInfo)
-	info, found = keeper.getValidatorSigningInfo(ctx, addrs[0])
+	keeper.setValidatorSigningInfo(ctx, sdk.ValAddress(addrs[0]), newInfo)
+	info, found = keeper.getValidatorSigningInfo(ctx, sdk.ValAddress(addrs[0]))
 	require.True(t, found)
 	require.Equal(t, info.StartHeight, int64(4))
 	require.Equal(t, info.IndexOffset, int64(3))
@@ -27,9 +29,9 @@ func TestGetSetValidatorSigningInfo(t *testing.T) {
 
 func TestGetSetValidatorSigningBitArray(t *testing.T) {
 	ctx, _, _, keeper := createTestInput(t)
-	signed := keeper.getValidatorSigningBitArray(ctx, addrs[0], 0)
+	signed := keeper.getValidatorSigningBitArray(ctx, sdk.ValAddress(addrs[0]), 0)
 	require.False(t, signed) // treat empty key as unsigned
-	keeper.setValidatorSigningBitArray(ctx, addrs[0], 0, true)
-	signed = keeper.getValidatorSigningBitArray(ctx, addrs[0], 0)
+	keeper.setValidatorSigningBitArray(ctx, sdk.ValAddress(addrs[0]), 0, true)
+	signed = keeper.getValidatorSigningBitArray(ctx, sdk.ValAddress(addrs[0]), 0)
 	require.True(t, signed) // now should be signed
 }