+++ x/stake/types/test_utils.go
@@ -14,11 +14,11 @@ var (
 	pk1   = crypto.GenPrivKeyEd25519().PubKey()
 	pk2   = crypto.GenPrivKeyEd25519().PubKey()
 	pk3   = crypto.GenPrivKeyEd25519().PubKey()
-	addr1 = pk1.Address()
-	addr2 = pk2.Address()
-	addr3 = pk3.Address()
+	addr1 = sdk.AccAddress(pk1.Address())
+	addr2 = sdk.AccAddress(pk2.Address())
+	addr3 = sdk.AccAddress(pk3.Address())
 
-	emptyAddr   sdk.Address
+	emptyAddr   sdk.AccAddress
 	emptyPubkey crypto.PubKey
 )